requirements.txt
pytest>=7.0
requests>=2.31
allure-pytest>=2.13